Ceph: add OSDs to crushmap with 0 weight
When you provision a large quantity of OSDs it takes some time to do all
the filesystem creation, ceph configuration and start all the services.
The inital crush weight should be 0 to avoid unecessary churn in the cluster
while all this is being done. Once the new OSDs are ready, running, up and in
we can then set the weight to 1 and then Ceph will only do a single map
calculation and start shuffling data.