Set node forced defaults for osd_devices elements
If you set osd_devices defaults on a role then when the recipe sets
a osd_device element's status to hold, we lose the rest of the elements
structure. By reading each of the values for the element and setting
them after a state change with node.force_default we can populate
osd_devices defaults in roles or on the node directly.